Added C payload containing all injection techniques
